+++ example/convex/_generated/api.d.ts
@@ -9,7 +9,7 @@
  */
 
 import type * as example from "../example.js";
-import type * as http from "../http.js";
+import type * as posthog from "../posthog.js";
 
 import type {
   ApiFromModules,
@@ -19,7 +19,7 @@ import type {
 
 declare const fullApi: ApiFromModules<{
   example: typeof example;
-  http: typeof http;
+  posthog: typeof posthog;
 }>;
 
 /**